+++ src/components/Tables/RelatedItemTable.vue
@@ -104,7 +104,7 @@
 
   const newItemLabel = computed(() => {
     switch (props.topic) {
-      case Topics.Event: return localize('labels.addTopic.event');
+      // case Topics.Event: return localize('labels.addTopic.event'); 
       case Topics.Character: return localize('labels.addTopic.character'); 
       case Topics.Location: return localize('labels.addTopic.location');
       case Topics.Organization: return localize('labels.addTopic.organization');
@@ -113,7 +113,7 @@
 
   const newItemDragLabel = computed(() => {
     switch (props.topic) {
-      case Topics.Event: return localize('labels.addTopicDrag.event');
+      // case Topics.Event: return localize('labels.addTopicDrag.event'); 
       case Topics.Character: return localize('labels.addTopicDrag.character'); 
       case Topics.Location: return localize('labels.addTopicDrag.location');
       case Topics.Organization: return localize('labels.addTopicDrag.organization');
@@ -154,11 +154,11 @@
     const dateColumn = { field: 'date', style: 'text-align: left', header: 'Date', format: (val: string) => (/*dateText(calendar.value, val)*/ val), sortable: true}; 
 
     const columns = {
-      [Topics.Event]: [
-        actionColumn,
-        nameColumn,
-        dateColumn,
-      ],
+      // [Topics.Event]: [ 
+      //   actionColumn,
+      //   nameColumn,
+      //   dateColumn,
+      // ],
       [Topics.Character]: [
         actionColumn,
         nameColumn,