+++ CHANGELOG.md
@@ -1,14 +1,15 @@
 # Change Log
 
-## v0.1.2 - [ IN PROGRESS ]
+## v0.1.2
 
-![](https://img.shields.io/badge/release%20date-April%2013%2C%202025-blue)
+![](https://img.shields.io/badge/release%20date-April%2020%2C%202025-blue)
 ![GitHub release](https://img.shields.io/github/downloads-pre/dovrosenberg/fvtt-campaign-builder/v0.1.2/module.zip)
 
 - Added buttons to add actors/scenes and drag indicator to scenes in Entry document relationship tabs
+- Support for Autocomplete Mentions module 
+- Made sessions searchable
 - Lots of documentation improvements
 - Various minor bug fixes
-
 
 ## v0.1.1 - Tags and quality of life